# backend/openapi.yaml
openapi: 3.1.0
info:
    title: EduFlow
    version: 0.1.0
    description: |
        EduFlow is a role-based education and workshop management backend API designed for Admin, Institutional Admin, Educator, Student, and Public users.

        The system supports the complete academic lifecycle including authentication, content delivery, assessments, certification, analytics, communication, and AI-powered features.

        #### Core capabilities include:

        - **Secure JWT-based authentication** and role-based access control for admin, institutional admin, educator, and student users.
        - **Institution and workshop administration**, including workshop creation, updates, enrollment tracking, assignments, and lifecycle management.
        - **Centralized content management** for modules, study materials, and learning resources uploaded by educators and accessed by students.
        - **Assessment management** for quizzes, tests, and assignments, including question creation, student attempts, answer submission, and result storage.
        - **Automated evaluation support** for objective assessments to reduce manual grading effort and improve consistency.
        - **Certificate generation** for successful workshop completion and public certificate verification using unique verification codes or identifiers.
        - **Performance tracking and reporting** for student progress, workshop statistics, assessment outcomes, completion rates, and trend analysis.
...

...

...

...

...

...

...

...

...

...

...

...

...
components:
    schemas:
...
        AssessmentResponse:
            properties:
                id:
                    title: Id
                    type: string
                module_id:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Module Id
                pass_mark:
                    anyOf:
                        - type: integer
                        - type: "null"
                    default: 0
                    title: Pass Mark
                title:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Title
                total_marks:
                    anyOf:
                        - type: integer
                        - type: "null"
                    title: Total Marks
                workshop_id:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Workshop Id
            required:
                - id
                - workshop_id
                - module_id
                - title
                - total_marks
            title: AssessmentResponse
            type: object
...
        Body_upload_material_api_v1_modules__module_id__materials_upload_post:
            properties:
                file:
                    contentMediaType: application/octet-stream
                    title: File
                    type: string
            required:
                - file
            title: Body_upload_material_api_v1_modules__module_id__materials_upload_post
            type: object
...
        HTTPValidationError:
            properties:
                detail:
                    items:
                        $ref: "#/components/schemas/ValidationError"
                    title: Detail
                    type: array
            title: HTTPValidationError
            type: object
...
        MaterialItem:
            properties:
                content:
                    title: Content
                    type: string
                created_at:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Created At
                id:
                    title: Id
                    type: string
                title:
                    title: Title
                    type: string
                type:
                    enum:
                        - video
                        - text
                        - link
                    title: Type
                    type: string
            required:
                - id
                - title
                - type
                - content
            title: MaterialItem
            type: object
...
        ModuleCreate:
            properties:
                materials:
                    default: []
                    items:
                        $ref: "#/components/schemas/MaterialItem"
                    title: Materials
                    type: array
                order_index:
                    default: 0
                    title: Order Index
                    type: integer
                title:
                    title: Title
                    type: string
                workshop_id:
                    title: Workshop Id
                    type: string
            required:
                - workshop_id
                - title
            title: ModuleCreate
            type: object
        ModuleReorderItem:
            description: Single entry in a batch reorder request.
            properties:
                module_id:
                    title: Module Id
                    type: string
                order_index:
                    title: Order Index
                    type: integer
            required:
                - module_id
                - order_index
            title: ModuleReorderItem
            type: object
        ModuleReorderRequest:
            description: "Body for PATCH /modules/reorder \u2014 batch set order_index."
            properties:
                items:
                    items:
                        $ref: "#/components/schemas/ModuleReorderItem"
                    title: Items
                    type: array
            required:
                - items
            title: ModuleReorderRequest
            type: object
        ModuleResponse:
            properties:
                id:
                    title: Id
                    type: string
                materials:
                    default: []
                    items:
                        $ref: "#/components/schemas/MaterialItem"
                    title: Materials
                    type: array
                order_index:
                    anyOf:
                        - type: integer
                        - type: "null"
                    title: Order Index
                title:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Title
                workshop_id:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Workshop Id
            required:
                - id
                - workshop_id
                - title
                - order_index
            title: ModuleResponse
            type: object
        ModuleUpdate:
            properties:
                materials:
                    anyOf:
                        - items:
                              $ref: "#/components/schemas/MaterialItem"
                          type: array
                        - type: "null"
                    title: Materials
                order_index:
                    anyOf:
                        - type: integer
                        - type: "null"
                    title: Order Index
                title:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Title
            title: ModuleUpdate
            type: object
...
        Page_AssessmentResponse_:
            properties:
                items:
                    items:
                        $ref: "#/components/schemas/AssessmentResponse"
                    title: Items
                    type: array
                limit:
                    title: Limit
                    type: integer
                offset:
                    title: Offset
                    type: integer
                total:
                    title: Total
                    type: integer
            required:
                - items
                - total
                - offset
                - limit
            title: Page[AssessmentResponse]
            type: object
...
        Page_ModuleResponse_:
            properties:
                items:
                    items:
                        $ref: "#/components/schemas/ModuleResponse"
                    title: Items
                    type: array
                limit:
                    title: Limit
                    type: integer
                offset:
                    title: Offset
                    type: integer
                total:
                    title: Total
                    type: integer
            required:
                - items
                - total
                - offset
                - limit
            title: Page[ModuleResponse]
            type: object
...
        ValidationError:
            properties:
                ctx:
                    title: Context
                    type: object
                input:
                    title: Input
                loc:
                    items:
                        anyOf:
                            - type: string
                            - type: integer
                    title: Location
                    type: array
                msg:
                    title: Message
                    type: string
                type:
                    title: Error Type
                    type: string
            required:
                - loc
                - msg
                - type
            title: ValidationError
            type: object
...
    securitySchemes:
        OAuth2PasswordBearer:
            flows:
                password:
                    scopes: {}
                    tokenUrl: /api/v1/auth/login
            type: oauth2
paths:
...
    /api/v1/assessments/module/{module_id}:
        get:
            operationId: list_by_module_api_v1_assessments_module__module_id__get
            parameters:
                - in: path
                  name: module_id
                  required: true
                  schema:
                      title: Module Id
                      type: string
                - description: Number of records to skip
                  in: query
                  name: offset
                  required: false
                  schema:
                      default: 0
                      description: Number of records to skip
                      minimum: 0
                      title: Offset
                      type: integer
                - description: Maximum records to return
                  in: query
                  name: limit
                  required: false
                  schema:
                      default: 50
                      description: Maximum records to return
                      maximum: 1000
                      minimum: 1
                      title: Limit
                      type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Page_AssessmentResponse_"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: List assessments for a module
            tags:
                - assessments
...
    /api/v1/modules/:
        post:
            operationId: create_api_v1_modules__post
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/ModuleCreate"
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ModuleResponse"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Create a module
            tags:
                - modules
    /api/v1/modules/reorder:
        patch:
            operationId: batch_reorder_api_v1_modules_reorder_patch
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/ModuleReorderRequest"
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/ModuleResponse"
                                title: Response Batch Reorder Api V1 Modules Reorder Patch
                                type: array
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Batch-reorder modules by setting their order_index values
            tags:
                - modules
    /api/v1/modules/{module_id}:
        delete:
            operationId: delete_one_api_v1_modules__module_id__delete
            parameters:
                - in: path
                  name: module_id
                  required: true
                  schema:
                      title: Module Id
                      type: string
            responses:
                "204":
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Delete a module
            tags:
                - modules
        get:
            operationId: get_one_api_v1_modules__module_id__get
            parameters:
                - in: path
                  name: module_id
                  required: true
                  schema:
                      title: Module Id
                      type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ModuleResponse"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Get a module by ID
            tags:
                - modules
        patch:
            operationId: update_one_api_v1_modules__module_id__patch
            parameters:
                - in: path
                  name: module_id
                  required: true
                  schema:
                      title: Module Id
                      type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/ModuleUpdate"
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ModuleResponse"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Update a module
            tags:
                - modules
    /api/v1/modules/{module_id}/materials/upload:
        post:
            operationId: upload_material_api_v1_modules__module_id__materials_upload_post
            parameters:
                - in: path
                  name: module_id
                  required: true
                  schema:
                      title: Module Id
                      type: string
            requestBody:
                content:
                    multipart/form-data:
                        schema:
                            $ref: "#/components/schemas/Body_upload_material_api_v1_modules__module_id__materials_upload_post"
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ModuleResponse"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Upload a material file and append it to the module's materials list
            tags:
                - modules
...
    /api/v1/workshops/{workshop_id}/modules:
        get:
            operationId: list_modules_api_v1_workshops__workshop_id__modules_get
            parameters:
                - in: path
                  name: workshop_id
                  required: true
                  schema:
                      title: Workshop Id
                      type: string
                - description: Number of records to skip
                  in: query
                  name: offset
                  required: false
                  schema:
                      default: 0
                      description: Number of records to skip
                      minimum: 0
                      title: Offset
                      type: integer
                - description: Maximum records to return
                  in: query
                  name: limit
                  required: false
                  schema:
                      default: 50
                      description: Maximum records to return
                      maximum: 1000
                      minimum: 1
                      title: Limit
                      type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Page_ModuleResponse_"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: List modules for a workshop
            tags:
                - workshops